Added a special_value variable, added a few lines to create_deck function, added more to the load_deck function
The special_char variable is meant to hold the names of the cards that do not have values ("reverse", "skip", etc.).
The load_deck function now successfully reads all of the values from deck.txt, the only problem is seperating the suits from the card value and assigning them to their proper location in the card struct.
Also added a simple psuedo-code for the mid-project submission, feel free to add anything you think is missing.